move table generation to its own file
