+++ urbanstats/export/export_csvs.py
@@ -1,7 +1,7 @@
 import fire
 import tqdm.auto as tqdm
 
-from create_website import shapefile_without_ordinals
+from urbanstats.website_data.table import shapefile_without_ordinals
 from urbanstats.statistics.output_statistics_metadata import (
     get_statistic_categories,
     statistic_internal_to_display_name,
