+++ urbanstats/games/quiz_custom.py
@@ -1,6 +1,6 @@
 from functools import lru_cache
 
-from create_website import shapefile_without_ordinals
+from urbanstats.website_data.table import shapefile_without_ordinals
 from urbanstats.statistics.output_statistics_metadata import (
     statistic_internal_to_display_name,
 )
